+++ v21_gfp_exact_sched_test/2processors/scheduler.h
@@ -1,1 +1,1 @@
-void sortTasksByPriorities(const state&, uint_fast8_t*);
+void sortTasksByPriorities(const state&, unsigned short*);
